+++ CHANGELOG.md
@@ -9,6 +9,17 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Add `NameRoles.IsEqual()`
+- Add `NameSet.IsEqual()`
+- Add `TypeRoles.IsEqual()`
+- Add `TypeSet.IsEqual()`
+- Add `EntityMessageNames.IsEqual()`
+- Add `EntityMessageTypes.IsEqual()`
+
 ## [0.1.1]
 
 ### Added